Downloaded Application Skeleton
I would far prefer to fork this than download it.  The download process for chrome on mac is seriously annoying.  The download window is a pop-under, so you never know where the darned thing went. It went into the downloads folder, and I have to do several clicks to get to the thing, unzip it, extract it, and get started on it.  Like I said, seriously annoying.